+++ frictionless/plugins/excel/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import ExcelPlugin
+from .dialect import ExcelDialect
+from .parser import XlsxParser, XlsParser
