+++ frictionless/plugins/excel/parser/__init__.py
@@ -1,0 +1,2 @@
+from .xls import XlsParser
+from .xlsx import XlsxParser
